requirements.txt
python-dotenv
openai
spotipy
ipywidgets
flask
Flask-Session